+++ controllers/volumegroupreplication_controller.go
@@ -22,6 +22,7 @@ import (
 	"time"
 
 	"github.com/go-logr/logr"
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"k8s.io/apimachinery/pkg/api/errors"
@@ -85,7 +86,7 @@ func (r *VolumeGroupReplicationReconciler) Reconcile(ctx context.Context, req ct
 	}
 
 	// Get VolumeGroupReplicationClass
-	_, err := v.getVolumeGroupReplicationClass(log, types.NamespacedName{
+	_, err := r.getVolumeGroupReplicationClass(log, types.NamespacedName{
 		Name:      v.instance.Spec.VolumeGroupReplicationClass,
 		Namespace: req.Namespace,
 	})
@@ -100,7 +101,7 @@ func (r *VolumeGroupReplicationReconciler) Reconcile(ctx context.Context, req ct
 		return ctrl.Result{}, nil
 	}
 
-	err = v.validatePVClabels()
+	err = r.validatePVClabels(v.instance, log)
 	if err != nil {
 		setFailureCondition(v.instance)
 
@@ -145,6 +146,71 @@ func (r *VolumeGroupReplicationReconciler) updateGroupReplicationStatus(
 		logger.Error(err, "failed to update status")
 
 		return err
+	}
+
+	return nil
+}
+
+type VGRInstance struct {
+	reconciler          *VolumeGroupReplicationReconciler
+	ctx                 context.Context
+	log                 logr.Logger
+	instance            *cachev1alpha1.VolumeGroupReplication
+	savedInstanceStatus cachev1alpha1.VolumeGroupReplicationStatus
+	namespacedName      string
+	result              ctrl.Result
+}
+
+// VolumeGroupReplicationReconciler get volume replication class object from the subjected namespace and return the same.
+func (r *VolumeGroupReplicationReconciler) getVolumeGroupReplicationClass(logger logr.Logger,
+	req types.NamespacedName) (*cachev1alpha1.VolumeGroupReplicationClass, error) {
+	vrcObj := &cachev1alpha1.VolumeGroupReplicationClass{}
+
+	err := r.Client.Get(context.TODO(), req, vrcObj)
+	if err != nil {
+		if errors.IsNotFound(err) {
+			logger.Error(err, "VolumeGroupReplicationClass not found", "VolumeGroupReplicationClass", req.Name)
+		} else {
+			logger.Error(err, "Got an unexpected error while fetching VolumeGroupReplicationClass", "VolumeGroupReplicationClass", req.Name)
+		}
+
+		return nil, err
+	}
+
+	return vrcObj, nil
+}
+
+func (r *VolumeGroupReplicationReconciler) validatePVClabels(
+	instance *cachev1alpha1.VolumeGroupReplication,
+	logger logr.Logger,
+) error {
+	pvcLabelSelector := instance.Spec.Selector
+
+	pvcSelector, err := metav1.LabelSelectorAsSelector(pvcLabelSelector)
+	if err != nil {
+		logger.Error(err, "error with PVC label selector", "pvcSelector", pvcLabelSelector)
+
+		return fmt.Errorf("error with PVC label selector, %w", err)
+	}
+	logger.Info("Fetching PersistentVolumeClaims", "pvcSelector", pvcLabelSelector)
+
+	listOptions := []client.ListOption{
+		client.MatchingLabelsSelector{
+			Selector: pvcSelector,
+		},
+	}
+
+	pvcList := &corev1.PersistentVolumeClaimList{}
+	if err := r.Client.List(context.TODO(), pvcList, listOptions...); err != nil {
+		logger.Error(err, "failed to list PersistentVolumeClaims", "pvcSelector", pvcLabelSelector)
+
+		return fmt.Errorf("failed to list PersistentVolumeClaims, %w", err)
+	}
+
+	logger.Info(fmt.Sprintf("Found %d PVCs using label selector %v", len(pvcList.Items), pvcLabelSelector))
+
+	for idx := range pvcList.Items {
+		logger.Info("PVC", "name", pvcList.Items[idx].Name, "namespace", pvcList.Items[idx].Namespace)
 	}
 
 	return nil
